Issues #162 feat: GifticonRecognizer 클래스 분리
