+++ data/src/main/java/com/lighthouse/util/recognizer/parser/BalanceParser.kt
@@ -1,0 +1,28 @@
+package com.lighthouse.util.recognizer.parser
+
+import com.lighthouse.util.recognizer.GifticonRecognizeInfo
+
+class BalanceParser {
+    private val cashCardFilterRegex = listOf(
+        CashCardRegex("(\\d{0,3}[,]?\\d{0,3}[,]\\d{1,3})".toRegex(), 1),
+        CashCardRegex("(\\d+)천원".toRegex(), 1000),
+        CashCardRegex("(\\d+)만원".toRegex(), 10000),
+        CashCardRegex("(\\d+)십만원".toRegex(), 100000)
+    )
+
+    fun parseCashCard(inputs: List<String>): BalanceParserResult {
+        val balance = inputs.firstNotNullOfOrNull { text ->
+            cashCardFilterRegex.firstNotNullOfOrNull { cashCard ->
+                val value = cashCard.regex.find(text)?.groupValues?.getOrNull(1)
+                value?.filter { it.isDigit() }?.toInt()?.times(cashCard.unit)
+            }
+        } ?: 0
+
+        return BalanceParserResult(isCashCard = balance > 0, balance = balance)
+    }
+
+    fun parseCashCard(info: GifticonRecognizeInfo): GifticonRecognizeInfo {
+        val result = parseCashCard(info.candidate)
+        return info.copy(isCashCard = result.isCashCard, balance = result.balance)
+    }
+}
