+++ data/src/main/java/com/lighthouse/util/recognizer/parser/BalanceParserResult.kt
@@ -1,0 +1,6 @@
+package com.lighthouse.util.recognizer.parser
+
+data class BalanceParserResult(
+    val isCashCard: Boolean,
+    val balance: Int
+)
