+++ data/src/main/java/com/lighthouse/util/recognizer/parser/BarcodeParserResult.kt
@@ -1,0 +1,6 @@
+package com.lighthouse.util.recognizer.parser
+
+data class BarcodeParserResult(
+    val barcode: String,
+    val filtered: List<String>
+)
