+++ data/src/main/java/com/lighthouse/util/recognizer/parser/ExpiredParserResult.kt
@@ -1,0 +1,8 @@
+package com.lighthouse.util.recognizer.parser
+
+import java.util.Date
+
+data class ExpiredParserResult(
+    val expired: Date,
+    val filtered: List<String>
+)
